Created plots script, added docstrings to signals groupping and plotting methods
